added package, moved classes into new directorys
