+++ src/Main.java
@@ -1,3 +1,6 @@
+import characters.subclasses.Allied;
+import characters.subclasses.Enemy;
+
 public class Main {
     public static void main(String[] args) throws Exception {
         // två objekt från vardera klasser
