+++ src/characters/Character.java
@@ -1,3 +1,5 @@
+package characters;
+
 public abstract class Character {
     
     protected String name;
